Simplify updateGameModel and its caller
For that, make clear that instantaneous events should NOT be mixed with
delayed ones. Otherwise the caller has to disambiguate them and
has to do a second call to updateGameModel hereby generating a second
list of delayed events; which cannot be merged with the first ones. The
right way is to recurse in updateGameModel, to apply instantaneous
events right away. Did that and documented it.